Don't use IncomingMessage#specific_from_name
For reasons [1] it would make life easier if we only showed the
authority name instead of the `IncomingMessage#specific_from_name` on
incoming message correspondence.

Currently we only use `specific_from_name?` for this display check, so
easy enough to override it to `false` so that we always fall back to the
authority name.

Fixes https://github.com/mysociety/wob-knop-theme/issues/26.

[1] https://groups.google.com/a/mysociety.org/g/alaveteli/c/SQtduOc_46Y/m/joEXtpUJAQAJ
[2] https://github.com/mysociety/alaveteli/search?q=specific_from_name%3F&unscoped_q=specific_from_name%3F